Create new dummy annotation

Use a real annotation to create a dummy annotation with degraded ratings and directly add info to the ChildRecordings class

New.annotations(
  row.meta,
  time.sd = NULL,
  change.cat.prob = NULL,
  new.name = NULL,
  ChildRecordings
)

Arguments

row.meta
: a row meta from ChildRecordings$all.meta to transform into dummy annotation

time.sd
: a time deviation in millisecond to apply to the segments on using a normal law

change.cat.prob
: The probability of a change in speaker_type

new.name
: a string containing a new annotator name

ChildRecordings
: a ChildRecordings class

Examples

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/vandam-daylong-demo"
CR = ChildRecordings(path)

#> ###############################################
#> Hello Wellcome to the ChildRecordings R Project 
#> 
#> Your ChildRecording project path contained : 
#>  8  annotations files 
#>  1  are referenced in the metadata 
#>  2  coders were found :  its vtc 
#> 
#> 4  file(s) seem(s) to be unreferenced in the metadata 
#> 	 more infos in  ChildRecordings$integrity_test$files.unreferenced 
#> 

time.sd = 300
change.cat.prob = 0.05
row.meta = CR$all.meta[2,]
new.name = "vtc_mod"
New.annotations(row.meta, time.sd, change.cat.prob, new.name, CR)