aggregation of annnotations data

Base on the result of an find.ratting.segment return or similar data.frame the function will extract data from annotation file in a raterData Class

# S3 method for rating
aggregate(data, ChildRecordings, cut = 0.1, verbose = T)

Arguments

data
: find.ratting.segment return or similar data.frame

ChildRecordings
: a ChildRecordings class

cut
: time size in second for the unit segment

Value

Soon

Details

the data will be organize un raw annotation format and a long segmented format

Examples

library(ChildRecordsR)
path = "A_childrecord_data_path"
CR = ChildRecordings(path)

#> Warning: cannot open file 'A_childrecord_data_pathmetadata/annotations.csv': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

# finding segments on wav file for designated rater
raters <- c("Coder1","Coder2","Coder3")
search <- find.ratting.segment(CR,"Wav_file_name", raters, range_from = t1, range_to = t2)

#> Error in find.ratting.segment(CR, "Wav_file_name", raters, range_from = t1,     range_to = t2): could not find function "find.ratting.segment"

ratting1  = aggregate.rating(search, CR, 0.1)

#> Error in is(ChildRecordings, "ChildRecordings"): object 'CR' not found

# try to analyse a larger number of file
wave_file <- unique(CR$all.meta$filename) # get all the wav files

#> Error in unique(CR$all.meta$filename): object 'CR' not found

ratters <- c("Coder1","Coder2","Coder3") # Define raters you are interested in

# bind all the results
search2 <- data.frame()
for (file in wave_file[1:10]){
  print(file)
  search2 <- rbind(search2, find.ratting.segment(CR, file, ratters)) # could take some time
}

#> Error in eval(expr, envir, enclos): object 'wave_file' not found

ratting2  = aggregate.rating(search2, CR, 0.1)

#> Error in is(ChildRecordings, "ChildRecordings"): object 'CR' not found