file importation from annotator

Import all the files from a specific annotator

extractDataCR(set.type, ChildRecordings, LENA.OL = F, verbose = T)

Arguments

set.type
: a string containing th name of the annotator present in the "set" column meta

ChildRecordings
: a ChildRecordings class

LENA.OL
: add the LENA overlap method. The overlap method will correct onset and offset when speech overlap by removing part that overlap.

Value

A class extractDataCR with a data.frame with all the aggregated data

Examples

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/vandam-daylong-demo"
CR = ChildRecordings(path)

#> ###############################################
#> Hello Wellcome to the ChildRecordings R Project 
#> 
#> Your ChildRecording project path contained : 
#>  9  annotations files 
#>  1  are referenced in the metadata 
#>  2  coders were found :  its vtc 
#> 
#> 5  file(s) seem(s) to be unreferenced in the metadata 
#> 	 more infos in  ChildRecordings$integrity_test$files.unreferenced 
#> 

rez = extractDataCR( "vtc", CR, verbose = F )
head(rez$data)

#>   child_id segment_onset segment_offset speaker_id ling_type speaker_type
#> 1        1          9992          10839         NA        NA       SPEECH
#> 2        1         10004          10814         NA        NA          CHI
#> 3        1         11298          11953         NA        NA       SPEECH
#> 4        1         11345          11828         NA        NA          CHI
#> 5        1         12113          12749         NA        NA          FEM
#> 6        1         12113          12753         NA        NA       SPEECH
#>   vcm_type lex_type mwu_type addresseee transcription phonemes syllables words
#> 1       NA       NA       NA         NA            NA       NA        NA    NA
#> 2       NA       NA       NA         NA            NA       NA        NA    NA
#> 3       NA       NA       NA         NA            NA       NA        NA    NA
#> 4       NA       NA       NA         NA            NA       NA        NA    NA
#> 5       NA       NA       NA         NA            NA       NA        NA    NA
#> 6       NA       NA       NA         NA            NA       NA        NA    NA
#>       raw_filename   date_iso range_offset range_onset  POSIXct_time_onset
#> 1 BN32_010007.rttm 2010-07-24     50464512           0 2010-07-24 14:44:32
#> 2 BN32_010007.rttm 2010-07-24     50464512           0 2010-07-24 14:44:44
#> 3 BN32_010007.rttm 2010-07-24     50464512           0 2010-07-24 15:06:18
#> 4 BN32_010007.rttm 2010-07-24     50464512           0 2010-07-24 15:07:05
#> 5 BN32_010007.rttm 2010-07-24     50464512           0 2010-07-24 15:19:53
#> 6 BN32_010007.rttm 2010-07-24     50464512           0 2010-07-24 15:19:53
#>   POSIXct_time_offset year month day     experiment  child_dob dob_criterion
#> 1 2010-07-24 14:58:39 2010     7  24 vandam-daylong 2009-07-24  extrapolated
#> 2 2010-07-24 14:58:14 2010     7  24 vandam-daylong 2009-07-24  extrapolated
#> 3 2010-07-24 15:17:13 2010     7  24 vandam-daylong 2009-07-24  extrapolated
#> 4 2010-07-24 15:15:08 2010     7  24 vandam-daylong 2009-07-24  extrapolated
#> 5 2010-07-24 15:30:29 2010     7  24 vandam-daylong 2009-07-24  extrapolated
#> 6 2010-07-24 15:30:33 2010     7  24 vandam-daylong 2009-07-24  extrapolated
#>   dob_accuracy age_in_day
#> 1        month        365
#> 2        month        365
#> 3        month        365
#> 4        month        365
#> 5        month        365
#> 6        month        365

# With LENA overlap method
rez = extractDataCR("vtc", CR, LENA.OL = T, verbose = F)