Openning a data file in childrecording class

Open a csv file from a ChildRecordings class. The function provide transformation on data such as create a true segment_onset and segment_offset based on meta data. Will also provide POSIXct time info if strat_time of the recording is provided in the meta

file.openner(meta_row, ChildRecordings)

Arguments

meta_row
: a row from the meta containing all the info of the cvs file

ChildRecordings
: a ChildRecordings class

Value

A data.frame

Examples

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)

#> ###############################################
#> Hello Wellcome to the ChildRecordings R Project 
#> 
#> Your ChildRecording project path containe : 
#>  1720  annotations files 
#>  1720  are referenced in the metadata 
#>  5  coders were finded :  textgrid_m1 vtc textgrid_mm textgrid_ak textgrid_ac
#>  All files seems to be present and referenced in the metadata (good news ;) )
#> 
#>   379  metadata don't have a start recording time (a.k.a start.time) 
#> 	 therefore time indicators wil not be build for those files 
#> 	 more infos in ChildRecordings$integrity_test$missing.start.time 
#>   362  files are empty 
#> 	 This should normally mean that no annotation were provided by annotator 
#> 	 more infos in ChildRecordings$integrity_test$empty.files 

raw_file <- file.openner(CR$all.meta[1,],CR)
head(raw_file)

#>   segment_onset segment_offset speaker_id ling_type speaker_type vcm_type
#> 1      1981.602       1982.215       CHI*         1          CHI       NA
#> 2      1982.215       1983.320         C1         1          OCH       NA
#> 3      1984.100       1985.210         C2         1          OCH       NA
#> 4      1985.593       1987.521       MOT*         1          FEM       NA
#> 5      1988.101       1989.285       LF2P         1         <NA>       NA
#> 6      1994.658       1996.058       CHI*         1          CHI       NA
#>   lex_type mwu_type addresseee transcription
#> 1       NA       NA         NA            NA
#> 2       NA       NA         NA            NA
#> 3       NA       NA         NA            NA
#> 4       NA       NA         NA            NA
#> 5       NA       NA         NA            NA
#> 6       NA       NA         NA            NA
#>                                        annotation_file   date_iso child_id
#> 1 textgrid/m1/namibie_aiku_20160714_1_1800_m1.TextGrid 2016-07-14     aiku
#> 2 textgrid/m1/namibie_aiku_20160714_1_1800_m1.TextGrid 2016-07-14     aiku
#> 3 textgrid/m1/namibie_aiku_20160714_1_1800_m1.TextGrid 2016-07-14     aiku
#> 4 textgrid/m1/namibie_aiku_20160714_1_1800_m1.TextGrid 2016-07-14     aiku
#> 5 textgrid/m1/namibie_aiku_20160714_1_1800_m1.TextGrid 2016-07-14     aiku
#> 6 textgrid/m1/namibie_aiku_20160714_1_1800_m1.TextGrid 2016-07-14     aiku
#>   range_offset range_onset  POSIXct_time_onset POSIXct_time_offset year month
#> 1          240         180 2016-07-14 09:44:01 2016-07-14 09:44:02 2016     7
#> 2          240         180 2016-07-14 09:44:02 2016-07-14 09:44:03 2016     7
#> 3          240         180 2016-07-14 09:44:04 2016-07-14 09:44:05 2016     7
#> 4          240         180 2016-07-14 09:44:05 2016-07-14 09:44:07 2016     7
#> 5          240         180 2016-07-14 09:44:08 2016-07-14 09:44:09 2016     7
#> 6          240         180 2016-07-14 09:44:14 2016-07-14 09:44:16 2016     7
#>   day                         filename
#> 1  14 aiku/namibie_aiku_20160714_1.wav
#> 2  14 aiku/namibie_aiku_20160714_1.wav
#> 3  14 aiku/namibie_aiku_20160714_1.wav
#> 4  14 aiku/namibie_aiku_20160714_1.wav
#> 5  14 aiku/namibie_aiku_20160714_1.wav
#> 6  14 aiku/namibie_aiku_20160714_1.wav