Search function on Childrecordings class

Search function in Childrecordings class This search function will help you to provide file name and time windows for common coding in raters

find.rating.segment(
  ChildRecordings,
  recording_filename,
  annotators = NULL,
  range_from = NULL,
  range_to = NULL
)

Arguments

ChildRecordings
: a ChildRecordings class

annotators
: an optional argument listing the annotators to consider

range_from
and range_to : an optional time window to restrain the search to

filename
: a wav file name to consider

Value

A data.frame containing sections that have been annotated by several annotations

Examples

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)

#> ###############################################
#> Hello Wellcome to the ChildRecordings R Project 
#> 
#> Your ChildRecording project path contained : 
#>  3619  annotations files 
#>  1559  are referenced in the metadata 
#>  7  coders were found :  textgrid/m1 alice_vtc vtc alice textgrid/ak textgrid/mm textgrid/ac 
#> 
#> 42  file(s) seem(s) to be unreferenced in the metadata 
#> 	 more infos in  ChildRecordings$integrity_test$files.unreferenced 
#> 
#>   427  metadata don't have a start recording time (a.k.a start.time) 
#> 	 therefore time indicators will not be built for those files 
#> 	 more infos in ChildRecordings$integrity_test$missing.start.time 
#>   362  files are empty 
#> 	 This should normally mean that no annotation were provided by annotator 
#> 	 more infos in ChildRecordings$integrity_test$empty.files 

# if no time windows is specified, this function will only return at table for all the know raters
# To work, all annotators must have at least one common annotation segment.
find.rating.segment(CR, "aiku/namibie_aiku_20160715_1.wav")

#>                 recording_filename         set
#> 1 aiku/namibie_aiku_20160715_1.wav textgrid/ak
#> 2 aiku/namibie_aiku_20160715_1.wav textgrid/m1
#> 3 aiku/namibie_aiku_20160715_1.wav   alice_vtc
#> 4 aiku/namibie_aiku_20160715_1.wav       alice
#> 5 aiku/namibie_aiku_20160715_1.wav         vtc
#> 6 aiku/namibie_aiku_20160715_1.wav textgrid/mm
#>                                annotation_filename true_onset true_offset
#> 1 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000
#> 2 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000
#> 3             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 4             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 5             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 6 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000

# However, if a time window is provided, this function will find all the data that
# overlaps with the time window provided.
# For instance, you can shift the window, and it will give you the same result
Wav_file_name = "aiku/namibie_aiku_20160715_1.wav"
t1 = 27180000
t2 = 27000000
t3 = 27240000
find.rating.segment(CR, Wav_file_name, range_from = t1, range_to = t3)

#>                 recording_filename         set
#> 1 aiku/namibie_aiku_20160715_1.wav textgrid/ak
#> 2 aiku/namibie_aiku_20160715_1.wav textgrid/m1
#> 3 aiku/namibie_aiku_20160715_1.wav   alice_vtc
#> 4 aiku/namibie_aiku_20160715_1.wav       alice
#> 5 aiku/namibie_aiku_20160715_1.wav         vtc
#> 6 aiku/namibie_aiku_20160715_1.wav textgrid/mm
#>                                annotation_filename true_onset true_offset
#> 1 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000
#> 2 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000
#> 3             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 4             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 5             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 6 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000

find.rating.segment(CR, Wav_file_name, range_from = t1, range_to = t3)

#>                 recording_filename         set
#> 1 aiku/namibie_aiku_20160715_1.wav textgrid/ak
#> 2 aiku/namibie_aiku_20160715_1.wav textgrid/m1
#> 3 aiku/namibie_aiku_20160715_1.wav   alice_vtc
#> 4 aiku/namibie_aiku_20160715_1.wav       alice
#> 5 aiku/namibie_aiku_20160715_1.wav         vtc
#> 6 aiku/namibie_aiku_20160715_1.wav textgrid/mm
#>                                annotation_filename true_onset true_offset
#> 1 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000
#> 2 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000
#> 3             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 4             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 5             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 6 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000

find.rating.segment(CR, Wav_file_name, range_from = t2, range_to = t3)

#>                 recording_filename         set
#> 1 aiku/namibie_aiku_20160715_1.wav textgrid/ak
#> 2 aiku/namibie_aiku_20160715_1.wav textgrid/m1
#> 3 aiku/namibie_aiku_20160715_1.wav   alice_vtc
#> 4 aiku/namibie_aiku_20160715_1.wav       alice
#> 5 aiku/namibie_aiku_20160715_1.wav         vtc
#> 6 aiku/namibie_aiku_20160715_1.wav textgrid/mm
#>                                annotation_filename true_onset true_offset
#> 1 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000
#> 2 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000
#> 3             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 4             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 5             aiku/namibie_aiku_20160715_1_0_0.csv   27180000    27240000
#> 6 aiku/namibie_aiku_20160715_1_27000000_180000.csv   27180000    27240000

# finding segments on wav file for designated rater
raters <-  c("textgrid_ak", "textgrid_mm", "textgrid_m1")
find.rating.segment(CR,"Wav_file_name", raters)

#> NULL

# finding segments on wav file for the designated windows in second and rater
search <- find.rating.segment(CR,"Wav_file_name", raters, range_from = t1, range_to = t2)

# try to analyse a larger number of file
wave_file <- unique(CR$all.meta$filename) # get all the wav files
raters <- c("textgrid_ak", "textgrid_mm", "textgrid_m1") # Define raters you are interested in

# bind all the results
search2 <- data.frame()
for (file in wave_file[1:10]){
  print(file)
  search2 <- rbind(search2, find.rating.segment(CR, file, raters)) # could take some time
}

head(search2)

#> data frame with 0 columns and 0 rows