Reliability analyse

Provide a reliability analyse returnin Krippen alpha, fleiss kappa and gwet AC1

reliability(raterData, summary = TRUE)

Arguments

raterData
: a raterData class

Value

Soon

Examples

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)

#> ###############################################
#> Hello Wellcome to the ChildRecordings R Project 
#> 
#> Your ChildRecording project path contained : 
#>  3619  annotations files 
#>  1559  are referenced in the metadata 
#>  7  coders were found :  textgrid/m1 alice_vtc vtc alice textgrid/ak textgrid/mm textgrid/ac 
#> 
#> 42  file(s) seem(s) to be unreferenced in the metadata 
#> 	 more infos in  ChildRecordings$integrity_test$files.unreferenced 
#> 
#>   427  metadata don't have a start recording time (a.k.a start.time) 
#> 	 therefore time indicators will not be built for those files 
#> 	 more infos in ChildRecordings$integrity_test$missing.start.time 
#>   362  files are empty 
#> 	 This should normally mean that no annotation were provided by annotator 
#> 	 more infos in ChildRecordings$integrity_test$empty.files 

# if no time windows is specified, this function will only return at table for all the know raters
# All the rater need to ratter any segment find
search = find.rating.segment(CR, "aiku/namibie_aiku_20160715_1.wav")
rez = aggregate.rating(search, CR, cut=0.100, verbose=T)

#> |============                                                              | 17% time by step : 0s estimate duration : 0m remain : 0m 
#> |=========================                                                 | 33% time by step : 0s estimate duration : 0m remain : 0m 
#> |=====================================                                     | 50% time by step : 0s estimate duration : 0m remain : 0m 
#> |=================================================                         | 67% time by step : 0s estimate duration : 0m remain : 0m 
#> |==============================================================            | 83% time by step : 0s estimate duration : 0m remain : 0m 
#> |==========================================================================|100% time by step : 0s estimate duration : 0m remain : 0m 
#> number of annotators 6 
#> length of recording annotation 360000 ms or  0.1 hours
#> Record span  60000 ms or  0.01666667 hours
#> 

reliability(rez)

#> Reliability indicator by type 
#> 
#> # CHI 
#> 
#>             coeff.name        pa        pe coeff.val coeff.se      conf.int
#> 1 Krippendorff's Alpha 0.8274506 0.7928866   0.16688  0.00050 (0.166,0.168)
#> 2        Fleiss' Kappa 0.8274505 0.7928866   0.16688  0.00050 (0.166,0.168)
#> 3                  AC1 0.8274505 0.2071134   0.78238  0.00047 (0.781,0.783)
#>   p.value     w.name
#> 1       0 unweighted
#> 2       0 unweighted
#> 3       0 unweighted
#> 
#> 
#> # OCH 
#> 
#>             coeff.name        pa       pe coeff.val coeff.se      conf.int
#> 1 Krippendorff's Alpha 0.7456276 0.603350    0.3587  0.00057  (0.358,0.36)
#> 2        Fleiss' Kappa 0.7456275 0.603350    0.3587  0.00057  (0.358,0.36)
#> 3                  AC1 0.7456275 0.198325    0.6827  0.00045 (0.682,0.684)
#>   p.value     w.name
#> 1       0 unweighted
#> 2       0 unweighted
#> 3       0 unweighted
#> 
#> 
#> # FEM 
#> 
#>             coeff.name        pa        pe coeff.val coeff.se      conf.int
#> 1 Krippendorff's Alpha 0.7475508 0.6008612   0.36752  0.00052 (0.366,0.369)
#> 2        Fleiss' Kappa 0.7475508 0.6008612   0.36752  0.00052 (0.366,0.369)
#> 3                  AC1 0.7475508 0.1995694   0.68461  0.00046 (0.684,0.686)
#>   p.value     w.name
#> 1       0 unweighted
#> 2       0 unweighted
#> 3       0 unweighted
#> 
#> 
#> # MAL 
#> 
#>             coeff.name        pa        pe coeff.val coeff.se      conf.int
#> 1 Krippendorff's Alpha 0.9304884 0.8927758   0.35172  0.00083  (0.35,0.353)
#> 2        Fleiss' Kappa 0.9304883 0.8927758   0.35172  0.00083  (0.35,0.353)
#> 3                  AC1 0.9304883 0.1072242   0.92214  0.00028 (0.922,0.923)
#>   p.value     w.name
#> 1       0 unweighted
#> 2       0 unweighted
#> 3       0 unweighted
#> 
#> 
#> # composit 
#> 
#>             coeff.name        pa        pe coeff.val coeff.se      conf.int
#> 1 Krippendorff's Alpha 0.4589163 0.2528483   0.27580  0.00034 (0.275,0.276)
#> 2        Fleiss' Kappa 0.4589161 0.2528483   0.27580  0.00034 (0.275,0.276)
#> 3                  AC1 0.4589161 0.1494303   0.36386  0.00042 (0.363,0.365)
#>   p.value     w.name
#> 1       0 unweighted
#> 2       0 unweighted
#> 3       0 unweighted
#> 
#>