aggregation of annnotations data

Base on the result of an find.ratting.segment return or similar data.frame the function will extract data from annotation file in a raterData Class

# S3 method for rating
aggregate(data, ChildRecordings, cut = 100, verbose = T)

Arguments

data
: find.ratting.segment return or similar data.frame

ChildRecordings
: a ChildRecordings class

cut
: time size in millisecond for the unit segment

Value

A raterData class containing with original format and long format for every annotators.

Details

the data will be organize un raw annotation format and a long segmented format

Examples

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)

#> ###############################################
#> Hello Wellcome to the ChildRecordings R Project 
#> 
#> Your ChildRecording project path contained : 
#>  3619  annotations files 
#>  1559  are referenced in the metadata 
#>  7  coders were found :  textgrid/m1 alice_vtc vtc alice textgrid/ak textgrid/mm textgrid/ac 
#> 
#> 42  file(s) seem(s) to be unreferenced in the metadata 
#> 	 more infos in  ChildRecordings$integrity_test$files.unreferenced 
#> 
#>   427  metadata don't have a start recording time (a.k.a start.time) 
#> 	 therefore time indicators will not be built for those files 
#> 	 more infos in ChildRecordings$integrity_test$missing.start.time 
#>   362  files are empty 
#> 	 This should normally mean that no annotation were provided by annotator 
#> 	 more infos in ChildRecordings$integrity_test$empty.files 

# if no time windows is specified, this function will only return at table for all the know raters
# All the rater need to ratter any segment find
search = find.rating.segment(CR, "aiku/namibie_aiku_20160715_1.wav")
rez = aggregate.rating(search, CR, cut=0.100, verbose=T)

#> |============                                                              | 17% time by step : 0s estimate duration : 0m remain : 0m 
#> |=========================                                                 | 33% time by step : 1s estimate duration : 0m remain : 0m 
#> |=====================================                                     | 50% time by step : 1s estimate duration : 0m remain : 0m 
#> |=================================================                         | 67% time by step : 0s estimate duration : 0m remain : 0m 
#> |==============================================================            | 83% time by step : 1s estimate duration : 0m remain : 0m 
#> |==========================================================================|100% time by step : 1s estimate duration : 0m remain : 0m 
#> number of annotators 6 
#> length of recording annotation for each annotator  60000 ms or  0.01666667 hours
#>