file importation from annotator

Import all the files from a specific annotator

extractDataCR(set.type, ChildRecordings, LENA.OL = F, verbose = T)

Arguments

set.type
: a string containing th name of the annotator present in the "set" column meta

ChildRecordings
: a ChildRecordings class

LENA.OL
: add the LENA overlap method. The overlap method will correct onset and offset when speech overlap by removing part that overlap.

Value

A class extractDataCR with a data.frame with all the aggregated data

Examples

library(ChildRecordsR)
path = "/mnt/94707AA4707A8CAC/CNRS/corpus/namibia-data/"
CR = ChildRecordings(path)

#> ###############################################
#> Hello Wellcome to the ChildRecordings R Project 
#> 
#> Your ChildRecording project path contained : 
#>  3619  annotations files 
#>  1559  are referenced in the metadata 
#>  7  coders were found :  textgrid/m1 alice_vtc vtc alice textgrid/ak textgrid/mm textgrid/ac 
#> 
#> 42  file(s) seem(s) to be unreferenced in the metadata 
#> 	 more infos in  ChildRecordings$integrity_test$files.unreferenced 
#> 
#>   427  metadata don't have a start recording time (a.k.a start.time) 
#> 	 therefore time indicators will not be built for those files 
#> 	 more infos in ChildRecordings$integrity_test$missing.start.time 
#>   362  files are empty 
#> 	 This should normally mean that no annotation were provided by annotator 
#> 	 more infos in ChildRecordings$integrity_test$empty.files 

rez = extractDataCR( "textgrid/m1", CR, verbose = F )
head(rez$data)

#>   child_id segment_onset segment_offset speaker_id ling_type speaker_type
#> 1     aiku      12780000       12780518       CHI*         1          CHI
#> 2     aiku      12781478       12783294         C1         1          OCH
#> 3     aiku      12783016       12783788       CHI*         1          CHI
#> 4     aiku      12784160       12786039       CHI*         1          CHI
#> 5     aiku      12786039       12786443       LF2P         1         <NA>
#> 6     aiku      12786420       12787016       CHI*         1          CHI
#>   vcm_type lex_type mwu_type addresseee transcription phonemes syllables words
#> 1       NA       NA       NA         NA            NA       NA        NA    NA
#> 2       NA       NA       NA         NA            NA       NA        NA    NA
#> 3       NA       NA       NA         NA            NA       NA        NA    NA
#> 4       NA       NA       NA         NA            NA       NA        NA    NA
#> 5       NA       NA       NA         NA            NA       NA        NA    NA
#> 6       NA       NA       NA         NA            NA       NA        NA    NA
#>                                raw_filename   date_iso range_offset range_onset
#> 1 namibie_aiku_20160714_1_12600_m1.TextGrid 2016-07-14       240000      180000
#> 2 namibie_aiku_20160714_1_12600_m1.TextGrid 2016-07-14       240000      180000
#> 3 namibie_aiku_20160714_1_12600_m1.TextGrid 2016-07-14       240000      180000
#> 4 namibie_aiku_20160714_1_12600_m1.TextGrid 2016-07-14       240000      180000
#> 5 namibie_aiku_20160714_1_12600_m1.TextGrid 2016-07-14       240000      180000
#> 6 namibie_aiku_20160714_1_12600_m1.TextGrid 2016-07-14       240000      180000
#>    POSIXct_time_onset POSIXct_time_offset year month day experiment child_sex
#> 1 2016-12-09 06:11:00 2016-12-09 06:19:38 2016    12   9    namibia         M
#> 2 2016-12-09 06:35:38 2016-12-09 07:05:54 2016    12   9    namibia         M
#> 3 2016-12-09 07:01:16 2016-12-09 07:14:08 2016    12   9    namibia         M
#> 4 2016-12-09 07:20:20 2016-12-09 07:51:39 2016    12   9    namibia         M
#> 5 2016-12-09 07:51:39 2016-12-09 07:58:23 2016    12   9    namibia         M
#> 6 2016-12-09 07:58:00 2016-12-09 08:07:56 2016    12   9    namibia         M
#>    child_dob mother_id father_id age_in_day
#> 1 2015-12-30         6         6        197
#> 2 2015-12-30         6         6        197
#> 3 2015-12-30         6         6        197
#> 4 2015-12-30         6         6        197
#> 5 2015-12-30         6         6        197
#> 6 2015-12-30         6         6        197

# With LENA overlap method
rez = extractDataCR("textgrid/m1", CR, LENA.OL = T, verbose = F)