Search function on Childrecordings class

Search function in Childrecordings class This search function will help you to provide file name and time windows for common coding in raters

find.rating.segment(
  ChildRecordings,
  filename,
  annotators = NULL,
  range_from = NULL,
  range_to = NULL
)

Arguments

ChildRecordings
: a ChildRecordings class

filename
: a wav file name to look

annotators
: an optional argument providing le annotators to look at

range_from
and range_to : an optional time windows for the search

Value

A data.frame containing commun annoation segment for each annotator

Examples

library(ChildRecordsR)
path = "A_childrecord_data_path"
CR = ChildRecordings(path)

#> Warning: cannot open file 'A_childrecord_data_pathmetadata/annotations.csv': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

CR

#> Error in eval(expr, envir, enclos): object 'CR' not found

# if no time windows is specified, this function will only return at table for all the know raters
# All the rater need to ratter any segment find
find.ratting.segment(CR,"Wav_file_name")

#> Error in find.ratting.segment(CR, "Wav_file_name"): could not find function "find.ratting.segment"

# However, if a time windows is provided, this function will find all the data that
# overlaps with the time windows provided.
# For instance, you can shift the window it will give you the same result
find.ratting.segment(CR,"Wav_file_name",range_from = t1, range_to = t3)

#> Error in find.ratting.segment(CR, "Wav_file_name", range_from = t1, range_to = t3): could not find function "find.ratting.segment"

find.ratting.segment(CR,"Wav_file_name",range_from = t1, range_to = t3)

#> Error in find.ratting.segment(CR, "Wav_file_name", range_from = t1, range_to = t3): could not find function "find.ratting.segment"

find.ratting.segment(CR,"Wav_file_name",range_from = t2, range_to = t3)

#> Error in find.ratting.segment(CR, "Wav_file_name", range_from = t2, range_to = t3): could not find function "find.ratting.segment"

# finding segments on wav file for designated rater
raters <- c("Coder1","Coder2","Coder3")
find.ratting.segment(CR,"Wav_file_name",raters)

#> Error in find.ratting.segment(CR, "Wav_file_name", raters): could not find function "find.ratting.segment"

# finding segments on wav file for the designated windows in second and rater
search <- find.ratting.segment(CR,"Wav_file_name", raters, range_from = t1, range_to = t2)

#> Error in find.ratting.segment(CR, "Wav_file_name", raters, range_from = t1,     range_to = t2): could not find function "find.ratting.segment"

# try to analyse a larger number of file
wave_file <- unique(CR$all.meta$filename) # get all the wav files

#> Error in unique(CR$all.meta$filename): object 'CR' not found

ratters <- c("Coder1","Coder2","Coder3") # Define raters you are interested in

# bind all the results
search2 <- data.frame()
for (file in wave_file[1:10]){
  print(file)
  search2 <- rbind(search2, find.ratting.segment(CR, file, ratters)) # could take some time
}

#> Error in eval(expr, envir, enclos): object 'wave_file' not found